access: abstracted par data manager checks for par membership.
